### 2017-12-06 #484 @arfon
👋 @sjspielman - it looks like you're `paper.md` file is referencing a bibtex file named `paper.bib` but there isn't a bibtex file with that name. Could you rename the bibtex file appropriately?
