### 2017-12-06 #484 @sjspielman
Yes, my apologies! It was accidentally named `bib.md`. I have now renamed it to `paper.bib` and pushed the change.
