### 2017-12-06 #484 @arfon
> Yes, my apologies! It was accidentally named bib.md. I have now renamed it to paper.bib and pushed the change.

Great, that's looking much better. This was just a 'pre-flight' check on my part to make sure the paper was compiling OK. Don't expect much activity until one of the JOSS topic editors has volunteered to handle this submission :-)
