### 2017-12-07 c111f0d
Reduced stringency of LEISR model, as this changes in HyPhy quite rapidly. Error should therefore be thrown by HyPhy, *not* phyphy
