### 2017-12-09 #484 @arfon
@brainstorm @biorelated @mgymrek @pjotrp - would one of your be willing to edit this submission for JOSS?
