### 2017-12-14 7873270
Updates to reflect LEISR v0.4, which is mulipartitioned and has more content output
