### 2017-12-17 #514 @whedon
[REVIEW]: phyphy: A Python package for facilitating the execution and parsing of HyPhy standard analyses
**Submitting author:** @sjspielman (<a href="http://orcid.org/0000-0002-9090-4788">Stephanie Spielman</a>)
**Repository:** <a href="https://github.com/sjspielman/phyphy" target ="_blank">https://github.com/sjspielman/phyphy</a>
**Version:** v0.4.1
