### 2018-01-08 #514 @mgymrek
Hi @luizirber I just wanted to check in on this. Please let me know if you have any questions.
