### 2018-01-10 #514 @sjspielman
Hello @mgymrek and @luizirber,

I have a inquiry regarding this review process. An updated version of one my citations has just been published and I would like to update my `paper.md` with this new citation. Is it possible to update at this stage or should I keep waiting until the review actually begins? It has now been about 4 weeks since the review was assigned so the time frame has not been clear to me.

Thank you,
Stephanie
