### 2018-01-10 05801ad
Updated paper for Datamonkey2 publication
