+++ paper/paper.bib
@@ -8,12 +8,10 @@
 	Year = {2005}
 }
 
-@article{datamonkey2010,
-	Author = {Delport, W. and Poon, A. F. and Frost, S. D. W. and {Kosakovsky Pond}, S. L.},
-	Journal = {Bioinformatics},
-	Pages = {2455--2457},
-	Title = {Datamonkey 2010: a suite of phylogenetic analysis tools for evolutionary biology},
-	Volume = {26},
-	Issue = {10},	
-	Year = {2010}
+@article{datamonkey2,
+	Author = {Weaver, S. and Shank, S. D. and Spielman, S. J. and Li, M. and Muse, S. V. and {Kosakovsky Pond}, S. L.},
+	Journal = {Molecular Biology and Evolution},
+	Pages = {msx335},
+	Title = {Datamonkey 2.0: a modern web application for characterizing selective and other evolutionary processes},
+	Year = {2018}
 }
