### 2018-01-10 #514 @sjspielman
Thanks very much! I just updated it. I realize, though, that it's still in press so the full citation with volume/issue/pages is still forthcoming. Either way it's better than before. Thanks again!
