### 2018-01-12 #514 @luizirber
Hello @sjspielman! Very nice software, just a couple of comments:

## Blockers
- [ ] DOI missing for datamonkey and HyPhy refs
  - [ ] HyPhy: https://doi.org/10.1093/bioinformatics/bti079
  - [ ] Datamonkey 2: https://doi.org/10.1093/molbev/msx335

## Recommendations

- LICENSE.txt doesn't appear to be recognized by github
  - It seems to be some oddness with detection the FreeBSD-style license, they recommend adding a note to README.md to detect it properly.
  - side note: any special reason to choose this license instead of a 3-clause or 2-clause BSD license?

- tests are a bit odd to run, any reason why they are python2.7 only?

- version listed on this review issue is 0.1, but latest release on github is 0.3 (but this can be fixed once there is a archived DOI)
