### 2018-01-13 c79d9b2
updated refs with doi
