+++ paper/paper.bib
@@ -5,7 +5,8 @@
 	Title = {{HyPhy}: hypothesis testing using phylogenetics},
 	Volume = {21},
 	Issue = {5},
-	Year = {2005}
+	Year = {2005},
+	doi = {https://doi.org/10.1093/bioinformatics/bti079}
 }
 
 @article{datamonkey2,
@@ -13,5 +14,6 @@
 	Journal = {Molecular Biology and Evolution},
 	Pages = {msx335},
 	Title = {Datamonkey 2.0: a modern web application for characterizing selective and other evolutionary processes},
-	Year = {2018}
+	Year = {2018},
+	doi = {https://doi.org/10.1093/molbev/msx335}
 }
