### 2018-01-13 #514 @sjspielman
HI @luizirber,

I have gone ahead and pushed the updated refs with the doi. For your other questions:

1) I have no good reason for using FreeBSD over the 2- or 3- clause, and if it helps github to recognize I will happily switch out.
2) The reason the tests are 2.7 compatible only has to do with changes to the unittest module in python, where certain functions that I need to use in testing were deprecated in python3. If you think this will cause a major issue, I can look into a 3rd party testing package, although I am hesitant to add further dependencies.
3) Yes, I did have to update the version with a few patches for some elusive bugs. Is there somewhere specific I should update this version information?

Thanks,
Stephanie
