### 2018-01-16 8ae4943
Addressed bug in specifiying non-comma delimitor for extract_csv
