### 2018-01-16 #514 @luizirber
Hi @sjspielman!

1. it's not a big issue (since the FreeBSD license is OSI approved), and start discussing software license is a neverending bikeshedding problem =P. GitHub uses https://github.com/benbalter/licensee/ to recognize licenses, but it seems to have problems with the FreeBSD one. I tend to use BSD 3-clause when I want a BSD license, but the final decision is up to you.
2. I think that's fine, although `pytest` is the de facto standard for Python testing nowadays.
3. The version is not listed in the paper, but once you mint a DOI for the archival copy (thru zenodo) it is important to have the correct version. @mgymrek is there a @whedon command to bump the version?

Other than that, if @sjspielman doesn't have any other corrections, checklist is completed and review is approved!
