### 2018-01-16 b2373af
Trying BSD3 license for github to recognize
