### 2018-01-16 #514 @sjspielman
Hi @luizirber,

Ok, I just went ahead to the BSD-3 clause and github finally recognizes it! I updated the `paper.md` accordingly, too.

In addition, in response to a user-identified bug, I updated the current release to 0.4 this morning, which can be final version for publishing. @mgymrek can hopefully advise on how to proceed with version bumping, as you mentioned.

Other than these, I'm all set on my end!

Thanks!
-Stephanie
