### 2018-01-17 #514 @mgymrek
Thanks @luizirber and @sjspielman! This looks good to me

I have not dealt with version bumps so will defer to @arfon:
since the review issue has been created the version of this software has been updated. Is there any @whedon command for that? I didn't see something obvious. Other than that it looks we're ready to accept this submission.
