+++ README.md
@@ -1,6 +1,8 @@
 # phyphy
 
-**The current release is version 0.4**.
+[![DOI](https://zenodo.org/badge/97511838.svg)](https://zenodo.org/badge/latestdoi/97511838)
+
+**The current release is version 0.4.1**.
 
 [`phyphy`](http://sjspielman.org/phyphy) aims to facilitate [HyPhy](http://hyphy.org) usage in two primary ways:
 1) Execute standard analyses in a Python scripting environment
