### 2018-01-17 #514 @sjspielman
Hi @arfon,

Got the DOI: [![DOI](https://zenodo.org/badge/97511838.svg)](https://zenodo.org/badge/latestdoi/97511838). I have also placed this badge in my repo's README.

However, it appeared that, in order to accomplish this archiving, I had to make another (sorry...) release of the repo. As such, the release tag associated with this DOI is 0.4.1.

Thanks,
Stephanie
